requirements.txt
requests==2.4.3
google-api-python-client==1.3.1
python-gflags==2.0
python-slugify==0.1.0
progressbar==2.2
